Move class Hyperlinks into Foundation\Kernel namespace
