+++ packages/framework/tests/Feature/Foundation/HyperlinksTest.php
@@ -5,22 +5,21 @@ declare(strict_types=1);
 namespace Hyde\Framework\Testing\Feature\Foundation;
 
 use Hyde\Foundation\HydeKernel;
-use Hyde\Foundation\Hyperlinks;
 use Hyde\Hyde;
 use Hyde\Testing\TestCase;
 
 /**
- * @covers \Hyde\Foundation\Hyperlinks
+ * @covers \Hyde\Foundation\Kernel\Hyperlinks
  */
 class HyperlinksTest extends TestCase
 {
-    protected Hyperlinks $class;
+    protected \Hyde\Foundation\Kernel\Hyperlinks $class;
 
     protected function setUp(): void
     {
         parent::setUp();
 
-        $this->class = new Hyperlinks(HydeKernel::getInstance());
+        $this->class = new \Hyde\Foundation\Kernel\Hyperlinks(HydeKernel::getInstance());
     }
 
     public function test_image_helper_gets_relative_web_link_to_image_stored_in_site_media_folder()
